Seperated ParleyEncryptedDataSource
